make sure we don't migrate twice

+++ resources/database/migrations/0000_00_00_000000_create_taggable_table.php
@@ -14,19 +14,23 @@ class CreateTaggableTable extends Migration
      */
     public function up()
     {
-        Schema::create('taggable_tags', function (Blueprint $table) {
-            $table->increments('tag_id');
-            $table->string('name');
-            $table->string('normalized');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('taggable_tags')) {
+            Schema::create('taggable_tags', function(Blueprint $table) {
+                $table->increments('tag_id');
+                $table->string('name');
+                $table->string('normalized');
+                $table->timestamps();
+            });
+        }
 
-        Schema::create('taggable_taggables', function (Blueprint $table) {
-            $table->unsignedInteger('tag_id');
-            $table->unsignedInteger('taggable_id');
-            $table->string('taggable_type');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('taggable_taggables')) {
+            Schema::create('taggable_taggables', function(Blueprint $table) {
+                $table->unsignedInteger('tag_id');
+                $table->unsignedInteger('taggable_id');
+                $table->string('taggable_type');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -36,7 +40,12 @@ class CreateTaggableTable extends Migration
      */
     public function down()
     {
-        Schema::drop('taggable_tags');
-        Schema::drop('taggable_taggables');
+        if (Schema::hasTable('taggable_tags')) {
+            Schema::drop('taggable_tags');
+        }
+
+        if (Schema::hasTable('taggable_taggables')) {
+            Schema::drop('taggable_taggables');
+        }
     }
 }
